fix: bound bundle cancellation cleanup
