+++ service/src/main/java/ai/floedb/floecat/service/query/catalog/ResolvedRelation.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2026 Yellowbrick Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ai.floedb.floecat.service.query.catalog;
+
+import ai.floedb.floecat.common.rpc.NameRef;
+import ai.floedb.floecat.common.rpc.QueryInput;
+import ai.floedb.floecat.common.rpc.ResourceId;
+import ai.floedb.floecat.metagraph.model.RelationNode;
+import ai.floedb.floecat.query.rpc.TableReferenceCandidate;
+
+/** One selected relation and the request-owned identity needed by pinning and payload assembly. */
+record ResolvedRelation(
+    TableReferenceCandidate candidate,
+    ResourceId relationId,
+    RelationNode node,
+    QueryInput selectedInput,
+    // Resolve once through the request memo so parallel builders never re-walk shared ancestors.
+    NameRef canonicalName) {}
